// client/src/pages/NewsFeed/types/Posts.ts
export type Posts ={
    id: number,
    name: string
    photo: string
    distance: string
    pace: string
    time: string
    medals_gold: number
    medals_silver: number
    medals_bronze: number
    date: string
    time_post: string
    description: string,
    like: number
    comments:number
    photo_post: string[],
}